freshrss/scripts/requirements.txt
requests>=2.28.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
